chore: readd api folder and cleanup the dummy ping API (#30)
Add back the `api` folder generated in the Apache HTTP client module
because otherwise the API interface like `NamespaceApi` is missing in
generated code.

Remove the ping API because that is a dummy API and is no longer useful
now we have the namespace APIs in.

Run `make gen` in rust to fill in the missing generated classes.