feat: add Java client and server framework (#69)
This PR introduces the basic client and server interactions expected in
Lance namespace:
1. any implementation must implement core's `LanceNamespace` interface
to support Java
2. I have added a basic `LanceRestNamespace` as example, but other
implementation like `LanceHiveNamespace` can also be provided as
separated modules.
3. I have implemented the SpringBoot controllers and a server as a Lance
REST adapter. The controllers take a `delegate`, which can be updated
later to be based on configs, for example if user configs `hive`, then
the server can initiate a `LanceHiveNamespace` and delegate executions
all to it. This will directly provides a REST server implementation for
Hive.