chore: leverage .openapi-generator-ignore in python client generation (#82)
This PR uses the `.openapi-generator-ignore` file for the Python client
generation, instead of running manual `rm` commands afterwards. The
OpenAPI generator for the Python client requires the ignore file to be
in the output directory.

However, the generation metadata `.openapi-generator/` is something that
can't be skipped in the python client generation, but there are some
open PRs to address this across all clients which seem to have been open
for a while now. For example:
https://github.com/OpenAPITools/openapi-generator/pull/15770.